Fix promiseValue of promiseRequest, and add generic promise handler
promiseRequest incorrectly assumed that the resolve and reject callbacks
where the callbacks passed through then and catch, and so assumed they
accept three arguments. Both resolve and reject accept only one
argument, which is the one set as promiseValue. PromiseHandlers had to
be updated to properly understand the promiseValues.

Add a generic declarative promise handler.